movieinfopage layout done
