+++ src/js/main.js
@@ -1,14 +1,14 @@
 import { checking } from "./navbar";
 import { carousel } from "./carousel";
 carousel();
-import {checking2} from './movieinfo.js';
-checking2();
+// import {checking2} from './movieinfo.js';
+// checking2();
 
 import { loadMovieAPI } from "./api.js";
 
 //console.log(loadMovieAPI());
 
-const movieLoad = await loadMovieAPI();
+// const movieLoad = await loadMovieAPI();
 //console.log("Movieload: "+movieLoad[0].title);
  //test();
