--- src/scss/indexPage/_movieinfo.scss
@@ -1,17 +1,0 @@
-#movieInfoWrapper {
-  background-color: $cDarkgrey;
-  color: $cDarkWhite;
-}
-#movieTitel {
-  text-align: center;
-}
-#movieInfoText {
-  text-align: center;
-}
-#movieBooking-cta {
-  background-color: $cDarkWhite;
-  padding: 12px;
-  font-weight: bold;
-  margin-top: 20rem;
-  margin-left: 15rem;
-}
